Complete refactor of logic
