+++ app/assets/src/components/views/bulk_download/BulkDownloadModalOptions.jsx
@@ -1,0 +1,322 @@
+import React from "react";
+import PropTypes from "~/components/utils/propTypes";
+import {
+  unset,
+  find,
+  filter,
+  get,
+  set,
+  some,
+  map,
+  isUndefined,
+  orderBy,
+  reject,
+} from "lodash/fp";
+import cx from "classnames";
+import memoize from "memoize-one";
+
+import StatusLabel from "~ui/labels/StatusLabel";
+import Dropdown from "~ui/controls/dropdowns/Dropdown";
+import LoadingMessage from "~/components/common/LoadingMessage";
+import RadioButton from "~ui/controls/RadioButton";
+import BasicPopup from "~/components/BasicPopup";
+import { createBulkDownload } from "~/api/bulk_downloads";
+import AccordionNotification from "~ui/notifications/AccordionNotification";
+import Notification from "~ui/notifications/Notification";
+import PrimaryButton from "~/components/ui/controls/buttons/PrimaryButton";
+import { UserContext } from "~/components/common/UserContext";
+
+import TaxonHitSelect from "./TaxonHitSelect";
+import cs from "./bulk_download_modal_options.scss";
+
+const triggersConditionalField = (conditionalField, selectedFields) =>
+  conditionalField.triggerValues.includes(
+    get(conditionalField.dependentField, selectedFields)
+  );
+
+class BulkDownloadModalOptions extends React.Component {
+  sortTaxaWithReadsOptions = memoize(options =>
+    orderBy(["sampleCount", "text"], ["desc", "asc"], options)
+  );
+
+  renderOption = (downloadType, field) => {
+    const {
+      backgroundOptions,
+      metricsOptions,
+      validSampleIds,
+      onFieldSelect,
+      selectedFields,
+      selectedDownloadTypeName,
+      conditionalFields,
+    } = this.props;
+
+    const selectedFieldsForType = get(selectedDownloadTypeName, selectedFields);
+    const selectedField = get(field.type, selectedFieldsForType);
+    let dropdownOptions = null;
+    let placeholder = "";
+    // Handle rendering conditional fields.
+
+    // For the file format field, render a placeholder. This is a special case.
+    const fileFormatConditionalField = conditionalFields[0];
+    if (
+      field.type === fileFormatConditionalField.field &&
+      downloadType.type === fileFormatConditionalField.downloadType &&
+      !triggersConditionalField(
+        fileFormatConditionalField,
+        selectedFieldsForType
+      )
+    ) {
+      return (
+        <div className={cs.field} key={field.type}>
+          <div className={cs.label}>{field.display_name}:</div>
+          <div className={cs.forcedOption}>.fasta</div>
+          <div className={cs.info}>
+            Note: Only .fasta is available when selecting one taxon.
+          </div>
+        </div>
+      );
+      // For other conditional fields, render nothing.
+    } else if (
+      conditionalFields.some(
+        conditionalField =>
+          field.type === conditionalField.field &&
+          downloadType.type === conditionalField.downloadType &&
+          !triggersConditionalField(conditionalField, selectedFieldsForType)
+      )
+    )
+      return;
+
+    // Set different props for the dropdown depending on the field type.
+    switch (field.type) {
+      case "file_format":
+        dropdownOptions = field.options.map(option => ({
+          text: option,
+          value: option,
+        }));
+
+        placeholder = "Select file format";
+        break;
+      case "taxa_with_reads":
+        return (
+          <div className={cs.field} key={field.type}>
+            <div className={cs.label}>{field.display_name}:</div>
+            <TaxonHitSelect
+              sampleIds={validSampleIds}
+              onChange={(value, displayName) => {
+                onFieldSelect(
+                  downloadType.type,
+                  field.type,
+                  value,
+                  displayName
+                );
+              }}
+              value={selectedField}
+              hitType="read"
+            />
+          </div>
+        );
+      case "taxa_with_contigs":
+        return (
+          <div className={cs.field} key={field.type}>
+            <div className={cs.label}>{field.display_name}:</div>
+            <TaxonHitSelect
+              sampleIds={validSampleIds}
+              onChange={(value, displayName) => {
+                onFieldSelect(
+                  downloadType.type,
+                  field.type,
+                  value,
+                  displayName
+                );
+              }}
+              value={selectedField}
+              hitType="contig"
+            />
+          </div>
+        );
+      case "background":
+        dropdownOptions = backgroundOptions || [];
+        placeholder = backgroundOptions ? "Select background" : "Loading...";
+        break;
+      case "metric":
+        dropdownOptions = metricsOptions || [];
+        placeholder = metricsOptions ? "Select metric" : "Loading...";
+        break;
+    }
+
+    if (!dropdownOptions) {
+      return null;
+    }
+
+    return (
+      <div className={cs.field} key={field.type}>
+        <div className={cs.label}>{field.display_name}:</div>
+        <Dropdown
+          fluid
+          placeholder={placeholder}
+          options={dropdownOptions}
+          onChange={(value, displayName) => {
+            onFieldSelect(downloadType.type, field.type, value, displayName);
+
+            // Reset conditional fields if they are no longer needed.
+            conditionalFields.forEach(conditionalField => {
+              if (
+                field.type === conditionalField.dependentField &&
+                downloadType.type === conditionalField.downloadType &&
+                !conditionalField.triggerValues.includes(value)
+              ) {
+                onFieldSelect(
+                  downloadType.type,
+                  conditionalField.field,
+                  undefined,
+                  undefined
+                );
+              }
+            });
+          }}
+          value={selectedField}
+          usePortal
+          withinModal
+        />
+      </div>
+    );
+  };
+
+  renderDownloadType = downloadType => {
+    const {
+      validSampleIds,
+      handleSelectDownloadType,
+      allSamplesUploadedByCurrentUser,
+      selectedDownloadTypeName,
+    } = this.props;
+    const { admin, appConfig } = this.context || {};
+
+    const selected = selectedDownloadTypeName === downloadType.type;
+    let disabled = false;
+    let disabledMessage = "";
+
+    if (
+      downloadType.uploader_only &&
+      !allSamplesUploadedByCurrentUser &&
+      !admin
+    ) {
+      disabled = true;
+      disabledMessage = `To download ${
+        downloadType.display_name
+      }, you must be the original uploader of all selected samples.`;
+    } else if (
+      downloadType.type === "original_input_file" &&
+      appConfig.maxSamplesBulkDownloadOriginalFiles &&
+      validSampleIds.size > appConfig.maxSamplesBulkDownloadOriginalFiles &&
+      !admin
+    ) {
+      disabled = true;
+      disabledMessage = `No more than ${
+        appConfig.maxSamplesBulkDownloadOriginalFiles
+      } samples
+        allowed for ${downloadType.display_name} downloads`;
+    }
+
+    const downloadTypeElement = (
+      <div
+        className={cx(
+          cs.downloadType,
+          selected && cs.selected,
+          disabled && cs.disabled
+        )}
+        key={downloadType.type}
+        onClick={() => !disabled && handleSelectDownloadType(downloadType.type)}
+      >
+        <RadioButton
+          disabled={disabled}
+          className={cs.radioButton}
+          selected={selected}
+        />
+        <div className={cs.content}>
+          <div className={cs.name}>
+            {downloadType.display_name}
+            {downloadType.file_type_display && (
+              <span className={cs.fileType}>
+                &nbsp;({downloadType.file_type_display})
+              </span>
+            )}
+            {downloadType.admin_only && (
+              <StatusLabel inline status="Admin Only" />
+            )}
+          </div>
+          <div className={cs.description}>{downloadType.description}</div>
+          {downloadType.fields &&
+            selected && (
+              <div className={cs.fields}>
+                {downloadType.fields.map(field =>
+                  this.renderOption(downloadType, field)
+                )}
+              </div>
+            )}
+        </div>
+      </div>
+    );
+
+    if (disabled && disabledMessage) {
+      return (
+        <BasicPopup
+          key={downloadType.type}
+          trigger={downloadTypeElement}
+          content={disabledMessage}
+        />
+      );
+    } else {
+      return downloadTypeElement;
+    }
+  };
+
+  renderDownloadTypes = () => {
+    const { downloadTypes } = this.props;
+
+    if (!downloadTypes) {
+      return <LoadingMessage message="Loading download types..." />;
+    }
+
+    const reportTypes = filter(["category", "report"], downloadTypes);
+    const rawTypes = filter(["category", "raw"], downloadTypes);
+
+    return (
+      <React.Fragment>
+        <div className={cs.category}>
+          <div className={cs.title}>Reports</div>
+          {reportTypes.map(this.renderDownloadType)}
+        </div>
+        <div className={cs.category}>
+          <div className={cs.title}>Raw Data</div>
+          {rawTypes.map(this.renderDownloadType)}
+        </div>
+      </React.Fragment>
+    );
+  };
+
+  render() {
+    return (
+      <div className={cs.downloadTypeContainer}>
+        {this.renderDownloadTypes()}
+      </div>
+    );
+  }
+}
+
+BulkDownloadModalOptions.propTypes = {
+  downloadTypes: PropTypes.arrayOf(PropTypes.DownloadType),
+  selectedDownloadTypeName: PropTypes.string,
+  // The selected fields of the currently selected download type.
+  selectedFields: PropTypes.objectOf(PropTypes.string),
+  onFieldSelect: PropTypes.func.isRequired,
+  validSampleIds: PropTypes.instanceOf(Set).isRequired,
+  backgroundOptions: PropTypes.array,
+  metricsOptions: PropTypes.array,
+  allSamplesUploadedByCurrentUser: PropTypes.bool,
+  handleSelectDownloadType: PropTypes.func.isRequired,
+  conditionalFields: PropTypes.arrayOf(PropTypes.object),
+};
+
+BulkDownloadModalOptions.contextType = UserContext;
+
+export default BulkDownloadModalOptions;
