+++ app/assets/src/components/views/bulk_download/bulk_download_modal.scss
@@ -1,0 +1,32 @@
+@import "~styles/themes/elements";
+@import "~styles/themes/typography";
+@import "~styles/themes/colors";
+
+.modal {
+  height: calc(85vh - 80px);
+  max-height: 1400px;
+  display: flex;
+  flex-direction: column;
+
+  .header {
+    .title {
+      @include font-header-xl;
+      // Reduce the amount of margin at the top of the modal.
+      margin-top: -$space-xs;
+    }
+
+    .tagline {
+      @include font-body-xs;
+      color: $medium-grey;
+    }
+  }
+
+  .options {
+    overflow: auto;
+  }
+
+  .footer {
+    margin-top: $space-m;
+    margin-bottom: $space-s;
+  }
+}
